fix signal/slot connection
svn path=/trunk/koffice/; revision=1029346

+++ krita/plugins/paintops/sumie/kis_sumi_shape_option.cpp
@@ -39,7 +39,7 @@ KisSumiShapeOption::KisSumiShapeOption()
     connect(m_options->oneDimBrushBtn,SIGNAL(toggled(bool)),SIGNAL( sigSettingChanged()));
     connect(m_options->twoDimBrushBtn,SIGNAL(toggled(bool)),SIGNAL( sigSettingChanged()));
     connect(m_options->mousePressureCBox,SIGNAL(toggled(bool)),SIGNAL( sigSettingChanged()));
-    connect(m_options->radiusSpinBox,SIGNAL(valueChanged(double)),SIGNAL( sigSettingChanged()));
+    connect(m_options->radiusSpinBox,SIGNAL(valueChanged(int)),SIGNAL( sigSettingChanged()));
     connect(m_options->sigmaSpinBox,SIGNAL(valueChanged(double)),SIGNAL( sigSettingChanged()));
     connect(m_options->rndBox,SIGNAL(valueChanged(double)),SIGNAL( sigSettingChanged()));
     connect(m_options->scaleBox,SIGNAL(valueChanged(double)),SIGNAL( sigSettingChanged()));
